requirements.txt
apipkg==1.5
asgiref==3.3.1
attrs==20.3.0
backcall==0.2.0
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
cssselect2==0.4.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.3
djangorestframework==3.12.2
drawSvg==1.7.0
drf-yasg==1.20.0
execnet==1.7.1
flake8==3.8.4
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
inflection==0.5.1
iniconfig==1.1.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
packaging==20.4
parso==0.7.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
pytest-cache==1.0
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-flake8==1.0.6
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
uWSGI==2.0.19.1
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.0
